# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaS API",
    "description": "Test and Simulation Enabler API. Generated from the request validation schemas the server enforces, so it cannot drift from the implementation.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A caller-safe failure message."
          },
          "details": {
            "type": "array",
            "description": "Per-field breakdown, carried by validation failures.",
            "items": {
              "type": "object",
              "properties": {
                "location": {
                  "type": "string",
                  "enum": ["params", "query", "body"]
                },
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": ["error"]
      }
    },
    "responses": {
      "ValidationError": {
        "description": "The request was malformed or failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No session, or it has expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The session may not do this (missing CSRF token, disallowed origin).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Too many requests from this client; retry later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unclassified server-side failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/simulation/start": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9][A-Za-z0-9 _\\-.()[\\]+@'#]*$",
                        "maxLength": 128
                      },
                      "devices": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "additionalProperties": false
                        }
                      },
                      "datasetId": {
                        "type": "string",
                        "maxLength": 256,
                        "nullable": true
                      },
                      "newDataset": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "maxLength": 256
                          },
                          "name": {
                            "type": "string",
                            "maxLength": 1024,
                            "nullable": true
                          },
                          "description": {
                            "type": "string",
                            "maxLength": 1024,
                            "nullable": true
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "maxLength": 256
                            }
                          },
                          "source": {
                            "type": "string",
                            "maxLength": 1024,
                            "nullable": true
                          }
                        },
                        "additionalProperties": true,
                        "nullable": true
                      },
                      "replayOptions": {
                        "type": "object",
                        "properties": {
                          "startTime": {
                            "type": "integer",
                            "minimum": 0,
                            "nullable": true
                          },
                          "endTime": {
                            "type": "integer",
                            "minimum": 0,
                            "nullable": true
                          },
                          "repeat": {
                            "type": "boolean"
                          },
                          "speedup": {
                            "type": "number",
                            "exclusiveMinimum": 0
                          }
                        },
                        "additionalProperties": true,
                        "nullable": true
                      },
                      "dataStorage": {
                        "type": "object",
                        "properties": {
                          "protocol": {
                            "type": "string",
                            "enum": ["MONGODB"]
                          },
                          "connConfig": {
                            "type": "object",
                            "properties": {
                              "host": {
                                "type": "string",
                                "pattern": "^[A-Za-z0-9._-]+$",
                                "maxLength": 253
                              },
                              "port": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 65535
                              },
                              "username": {
                                "type": "string",
                                "maxLength": 256,
                                "nullable": true
                              },
                              "password": {
                                "type": "string",
                                "maxLength": 256,
                                "nullable": true
                              },
                              "dbname": {
                                "type": "string",
                                "maxLength": 256,
                                "nullable": true
                              },
                              "options": {
                                "anyOf": [
                                  {
                                    "type": "object",
                                    "additionalProperties": false
                                  },
                                  {
                                    "type": "string",
                                    "maxLength": 2048
                                  }
                                ]
                              }
                            },
                            "required": ["host", "port"],
                            "additionalProperties": true
                          }
                        },
                        "required": ["connConfig", "protocol"],
                        "additionalProperties": true,
                        "nullable": true
                      },
                      "testCampaignId": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9][A-Za-z0-9 _\\-.()[\\]+@'#]*$",
                        "maxLength": 128,
                        "nullable": true
                      },
                      "evaluationParameters": {
                        "type": "object",
                        "properties": {
                          "threshold": {
                            "type": "number"
                          },
                          "eventType": {
                            "type": "string",
                            "maxLength": 256
                          },
                          "metricType": {
                            "type": "string",
                            "maxLength": 256
                          }
                        },
                        "additionalProperties": true,
                        "nullable": true
                      }
                    },
                    "required": ["devices", "name"],
                    "additionalProperties": true
                  },
                  "modelFileName": {
                    "type": "string",
                    "maxLength": 133
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "datasetId": {
                        "type": "string",
                        "maxLength": 256,
                        "nullable": true
                      },
                      "newDataset": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "maxLength": 256
                          },
                          "name": {
                            "type": "string",
                            "maxLength": 1024,
                            "nullable": true
                          },
                          "description": {
                            "type": "string",
                            "maxLength": 1024,
                            "nullable": true
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "maxLength": 256
                            }
                          },
                          "source": {
                            "type": "string",
                            "maxLength": 1024,
                            "nullable": true
                          }
                        },
                        "additionalProperties": true,
                        "nullable": true
                      },
                      "replayOptions": {
                        "type": "object",
                        "properties": {
                          "startTime": {
                            "type": "integer",
                            "minimum": 0,
                            "nullable": true
                          },
                          "endTime": {
                            "type": "integer",
                            "minimum": 0,
                            "nullable": true
                          },
                          "repeat": {
                            "type": "boolean"
                          },
                          "speedup": {
                            "type": "number",
                            "exclusiveMinimum": 0
                          }
                        },
                        "additionalProperties": true,
                        "nullable": true
                      },
                      "dataStorage": {
                        "type": "object",
                        "properties": {
                          "protocol": {
                            "type": "string",
                            "enum": ["MONGODB"]
                          },
                          "connConfig": {
                            "type": "object",
                            "properties": {
                              "host": {
                                "type": "string",
                                "pattern": "^[A-Za-z0-9._-]+$",
                                "maxLength": 253
                              },
                              "port": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 65535
                              },
                              "username": {
                                "type": "string",
                                "maxLength": 256,
                                "nullable": true
                              },
                              "password": {
                                "type": "string",
                                "maxLength": 256,
                                "nullable": true
                              },
                              "dbname": {
                                "type": "string",
                                "maxLength": 256,
                                "nullable": true
                              },
                              "options": {
                                "anyOf": [
                                  {
                                    "type": "object",
                                    "additionalProperties": false
                                  },
                                  {
                                    "type": "string",
                                    "maxLength": 2048
                                  }
                                ]
                              }
                            },
                            "required": ["host", "port"],
                            "additionalProperties": true
                          }
                        },
                        "required": ["connConfig", "protocol"],
                        "additionalProperties": true,
                        "nullable": true
                      },
                      "testCampaignId": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9][A-Za-z0-9 _\\-.()[\\]+@'#]*$",
                        "maxLength": 128,
                        "nullable": true
                      },
                      "evaluationParameters": {
                        "type": "object",
                        "properties": {
                          "threshold": {
                            "type": "number"
                          },
                          "eventType": {
                            "type": "string",
                            "maxLength": 256
                          },
                          "metricType": {
                            "type": "string",
                            "maxLength": 256
                          }
                        },
                        "additionalProperties": true,
                        "nullable": true
                      }
                    },
                    "additionalProperties": false,
                    "default": {}
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/simulation/stop/{fileName}": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9 _\\-.()\\[\\]+@'#]*\\.json$",
              "maxLength": 133
            }
          }
        ]
      }
    },
    "/api/simulation/status": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/simulation/stats": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}